+++ lib/ListOfSps.php
@@ -8,20 +8,23 @@ class ListOfSps
         return strcmp(strtolower($a['facility']->getName()), strtolower($b['facility']->getName()));
     }
 
-    public static function getClass($attribute)
+    public static function getClass($type)
     {
-        if ($attribute['type'] === 'java.lang.String' || $attribute['type'] === 'java.lang.LargeString') {
-            return 'string';
-        } elseif ($attribute['type'] === 'java.lang.Integer') {
-            return 'integer';
-        } elseif ($attribute['type'] === 'java.lang.Boolean') {
-            return 'boolean';
-        } elseif ($attribute['type'] === 'java.util.ArrayList' || $attribute['type'] === 'java.util.LargeArrayList') {
-            return 'array';
-        } elseif ($attribute['type'] === 'java.util.LinkedHashMap') {
-            return 'map';
-        } else {
-            return '';
+        switch ($type) {
+            case 'java.lang.String':
+            case 'java.lang.LargeString':
+                return 'string';
+            case 'java.lang.Integer':
+                return 'integer';
+            case 'java.lang.Boolean':
+                return 'boolean';
+            case 'java.util.ArrayList':
+            case 'java.util.LargeArrayList':
+                return 'array';
+            case 'java.util.LinkedHashMap':
+                return 'map';
+            default:
+                return '';
         }
     }
 
@@ -34,44 +37,77 @@ class ListOfSps
         return "<a class='customLink' href='" . htmlspecialchars($loginURL) . "'>" . htmlspecialchars($name) . "</a>";
     }
 
-    public static function printAttributeValue($attribute, $service, $attr)
+    public static function printAttributeValue($type, $value)
     {
-        $value = $attribute['value'];
-        if (empty($value) && $attribute['type'] !== 'java.lang.Boolean') {
+        if (empty($value) && $type !== 'java.lang.Boolean') {
             return "<td class='center'>&horbar;</td>";
         }
-        $string = '';
-        if ($attribute['type'] === 'java.lang.String' || $attribute['type'] === 'java.lang.LargeString') {
-            if (filter_var($value, FILTER_VALIDATE_URL)) {
-                $string = '<a class="customLink" href="' . $value . '">' . $value . '</a>';
-            } else {
-                $string = $value;
-            }
-        } elseif ($attribute['type'] === 'java.lang.Integer') {
-            $string = $value;
-        } elseif ($attribute['type'] === 'java.lang.Boolean') {
-            if ($value !== null && $value) {
-                $string = '&#x2714;';
-            } else {
-                $string = '&#x2715;';
-            }
-        } elseif ($attribute['type'] === 'java.util.ArrayList' || $attribute['type'] === 'java.lang.LargeArrayList') {
-            $string = '<ul>';
-            foreach ($value as $v) {
-                $string .= '<li>' . $v . '</li>';
-            }
-            $string .= '</ul>';
-        } elseif ($attribute['type'] === 'java.util.LinkedHashMap') {
-            $string = '<ul>';
-            foreach ($value as $k => $v) {
-                $string .= '<li>' . $k . ' &rarr; ' . $v . '</li>';
-            }
-            $string .= '</ul>';
+
+        switch ($type) {
+            case 'java.lang.String':
+            case 'java.lang.LargeString':
+                if (filter_var($value, FILTER_VALIDATE_URL)) {
+                    $string = '<a class="customLink" href="' . htmlspecialchars($value) . '">'
+                        . htmlspecialchars($value) . '</a>';
+                } else {
+                    $string = htmlspecialchars($value);
+                }
+                break;
+            case 'java.lang.Integer':
+                $string = htmlspecialchars($value);
+                break;
+            case 'java.lang.Boolean':
+                if ($value !== null && $value) {
+                    $string = '&#x2714;';
+                } else {
+                    $string = '&#x2715;';
+                }
+                break;
+            case 'java.util.ArrayList':
+            case 'java.lang.LargeArrayList':
+                $string = '<ul>';
+                foreach ($value as $v) {
+                    $string .= '<li>' . htmlspecialchars($v) . '</li>';
+                }
+                $string .= '</ul>';
+                break;
+            case 'java.util.LinkedHashMap':
+                $string = '<ul>';
+                foreach ($value as $k => $v) {
+                    $string .= '<li>' . htmlspecialchars($k) . ' &rarr; ' . htmlspecialchars($v) . '</li>';
+                }
+                $string .= '</ul>';
+                break;
+            default:
+                $string = '';
         }
         if (!empty($string)) {
-            return '<td class="' . self::getClass($service['facilityAttributes'][$attr]) . '">' . $string . '</td>';
+            return '<td class="' . self::getClass($type) . '">' . $string . '</td>';
         } else {
             return '<td/>';
         }
+    }
+
+    public static function getPreferredTranslation($translations, $language = 'en')
+    {
+        if (is_string($translations)) {
+            return $translations;
+        }
+
+        if (isset($translations[$language])) {
+            return $translations[$language];
+        }
+
+        if (isset($translations['en'])) {
+            return $translations['en'];
+        }
+
+        if (count($translations) > 0) {
+            $languages = array_keys($translations);
+            return $translations[$languages[0]];
+        }
+
+        // we don't have anything to return
+        throw new \Exception('Nothing to return from translation.');
     }
 }